feat: added display abstraction layer
